feat: adopt useCumulativeCollateralVolumes
