+++ src/pages/Dashboard/sub-pages/Home/LockedCollateralsCard/index.tsx
@@ -1,52 +1,34 @@
 import { useSelector } from 'react-redux';
 import { useTranslation } from 'react-i18next';
-import { useQuery } from 'react-query';
 import { useErrorHandler, withErrorBoundary } from 'react-error-boundary';
-import { CollateralUnit } from '@interlay/interbtc-api';
 
 import LineChart from '../../../LineChart';
 import DashboardCard from '../../../cards/DashboardCard';
 import Stats, { StatsDt, StatsDd, StatsRouterLink } from '../../../Stats';
 import ErrorFallback from 'components/ErrorFallback';
-import { RELAY_CHAIN_NATIVE_TOKEN_SYMBOL, RELAY_CHAIN_NATIVE_TOKEN, WRAPPED_TOKEN } from 'config/relay-chains';
+import { RELAY_CHAIN_NATIVE_TOKEN_SYMBOL, RELAY_CHAIN_NATIVE_TOKEN } from 'config/relay-chains';
 import { POLKADOT, KUSAMA } from 'utils/constants/relay-chain-names';
 import { INTERLAY_DENIM, KINTSUGI_SUPERNOVA } from 'utils/constants/colors';
 import { PAGES } from 'utils/constants/links';
-import { getUsdAmount, displayMonetaryAmount, getLastMidnightTimestamps } from 'common/utils/utils';
+import { getUsdAmount, displayMonetaryAmount } from 'common/utils/utils';
 import { StoreType } from 'common/types/util.types';
-import cumulativeVolumesFetcher, {
-  CUMULATIVE_VOLUMES_FETCHER,
-  VolumeDataPoint,
-  VolumeType
-} from 'services/fetchers/cumulative-volumes-fetcher';
-
-// get 6 values to be able to calculate difference between 5 days ago and 6 days ago
-// thus issues per day 5 days ago can be displayed
-// cumulative issues is also only displayed to 5 days
-const cutoffTimestamps = getLastMidnightTimestamps(6, true);
+// ray test touch <
+import useCumulativeCollateralVolumes from 'services/hooks/use-cumulative-collateral-volumes';
+// ray test touch >
 
 const LockedCollateralsCard = (): JSX.Element => {
   const { prices } = useSelector((state: StoreType) => state.general);
   const { t } = useTranslation();
 
+  // ray test touch <
   const {
     isIdle: cumulativeVolumesIdle,
     isLoading: cumulativeVolumesLoading,
     data: cumulativeVolumes,
     error: cumulativeVolumesError
-    // TODO: should type properly (`Relay`)
-  } = useQuery<VolumeDataPoint<CollateralUnit>[], Error>(
-    [
-      CUMULATIVE_VOLUMES_FETCHER,
-      VolumeType.Collateral,
-      cutoffTimestamps,
-      RELAY_CHAIN_NATIVE_TOKEN, // returned amounts
-      RELAY_CHAIN_NATIVE_TOKEN, // filter by this collateral...
-      WRAPPED_TOKEN // and this backing currency
-    ],
-    cumulativeVolumesFetcher
-  );
+  } = useCumulativeCollateralVolumes(RELAY_CHAIN_NATIVE_TOKEN);
   useErrorHandler(cumulativeVolumesError);
+  // ray test touch >
 
   const renderContent = () => {
     // TODO: should use skeleton loaders
